- Basic alien enemy speed & animation speed system set up - Map adjustments made

+++ Assets/Scripts/playerController.cs
@@ -14,7 +14,7 @@ public class playerController : MonoBehaviour
     [SerializeField] float jumpHeight;
     [SerializeField] float gravityValue;
     [SerializeField] int jumpCount;
-    [SerializeField] int sprintMod;
+    [SerializeField] float sprintMod;
     [SerializeField] int dashMod;
     [SerializeField] float dashTime;
